motoactions: Improve doze denial thresholds
Previously we would deny a doze if it had been more than 1.5 seconds
since the last doze started.  A doze lasts about 7 seconds so that
was a meaningless test except as it related to the first doze after
the screen turned off.

Break that constant into two constants (time since the screen turned
off and time since the last doze started) to allow them to be
independently configured.

Make the screen on delay longer (3 seconds) to avoid excessively
dozing as you turn the screen off.

Make the inter-doze time 10 seconds which seems to be a balance between
waving your hand fruitlessly toward your phone and triggering too
easily.

Tihs is a hacky fix to the fact that the stow sensor seems to
erroneously
trigger a false followed short by a true if stowed when the doze
finishes.

Fixes Lineage: BUGBASH-42